requirements.txt
alabaster==0.7.16
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autopep8==2.0.4
Babel==2.14.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
config==0.5.1
ConfigArgParse==1.7
coverage==7.4.1
cryptography==41.0.7
defusedxml==0.8.0rc2
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.1
django-appconf==1.0.6
django-autocomplete-light==3.9.7
django-compressor==4.4
django-debug-toolbar==4.2.0
django-filter==23.5
django-htmlmin==0.11.0
django-parler==2.3
django-parler-rest==2.2
django-querysetsequence==0.17
django-redis==5.4.0
django-rosetta==0.9.8
django-silk==5.1.0
django-solo==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
docutils==0.20.1
drf-nested-routers==0.93.5
drf-yasg==1.21.7
filelock==3.13.1
Flask==3.0.1
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
flower==2.0.1
gevent==23.9.1
geventhttpclient==2.0.11
gprof2dot==2022.7.29
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
humanize==4.9.0
idna==3.6
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
locust==2.20.1
lxml==5.1.0
MarkupSafe==2.1.4
model-bakery==1.17.0
msgpack==1.0.7
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.3.0
polib==1.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pymemcache==4.0.0
pyproject-api==1.6.1
PySocks==1.7.1
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-memcached==1.62
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rcssmin==1.1.1
redis==5.0.1
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rjsmin==1.2.1
roundrobin==0.0.4
selenium==4.18.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
tornado==6.4
tox==4.12.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
wsproto==1.2.0
zeep==4.2.1
zope.event==5.0
zope.interface==6.1
